* Entity is defined as uint64_t

+++ EngineCore/Common/Archive.h
@@ -5,7 +5,7 @@
 
 #include <string>
 
-using Entity = uint32_t;
+using Entity = uint64_t;
 
 namespace vz
 {
